[ADD] dynamiquely draw range of turrets

+++ includes/structs.h
@@ -71,7 +71,7 @@ typedef struct turret_s {
     sfVector2f position;
     int damage_speed;
     int damage_per_action;
-    int range;
+    float range;
     struct turret_s *next;
 } turret_t;
